% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cellr_morgenstern_et_al_2017}
\alias{cellr_morgenstern_et_al_2017}
\title{Morgenstern et al., Cell Reports 2017}
\format{A data frame with 1576 rows and 54 variables}
\usage{
data(cellr_morgenstern_et_al_2017)
}
\description{
Project description: Mitochondria perform central functions in cellular bioenergetics, metabolism and signaling and their malfunction has been linked to numerous diseases.
The available studies cover only part of the mitochondrial proteome and a separation of core mitochondrial proteins from associated fractions has not been achieved.
We developed an integrative, quantitative MS-based experimental approach to define the high confidence proteome of yeast mitochondria and to identify new mitochondrial proteins.
The analysis includes protein abundance under fermentable and non-fermentable growth, submitochondrial localization, single-protein analysis and subcellular classification of mitochondria-associated fractions.
We identified novel mitochondrial interactors of respiratory chain supercomplexes, ATP synthase, AAA proteases, the mitochondrial contact site and cristae organizing system (MICOS) and coenzyme Q biosynthesis cluster as well as new mitochondrial proteins with dual cellular localization.
The integrative proteome provides a high confidence source for the characterization of physiological and pathophysiological functions of mitochondria and their integration into the cellular environment.
Data: This data are a summary of proteome-wide absolute quantification of proteins under fermentable and non-fermentable growth conditions based on the 'Proteomic Ruler' approach.
Raw MS data and complete MaxQuant results files are available via ProteomeXchange with identifier PXD006146.
The variables are as follows:
}
\details{
\itemize{
\item Systematic_names systematic name, Saccharomyces Genome Datavase (SGD)
\item Uniprot_ID Universal Protein database (UniProt) identifier
\item Gene_names name of gene
\item Protein_Description protein description
\item Moleular_weight_Da molecular weight of the leading protein sequence
\item High_confidence_mito_proteome reported mitochondria proteome
\item Sequence_coverage coverage of protein sequence by the identified peptides
\item Mean_copy_number_Glucose protein mean copy number in cells grown on glucose
\item Mean_copy_number_Galactose protein mean copy number in cells grown on galactose
\item Mean_copy_number_Glycerol protein mean copy number in cells grown on glycerol
\item Absolute_quantification_accuracy degree of quantification accuracy, according to Perseus
\item Copy_number_Kulak_et_al_2014 protein copy number reported by Kulak et al 2014
\item Copy_number_Chong_et_al_2015 protein copy number reported by Chong et al 2015
\item Copy_number_Ghaemmaghami_et_al_2003 protein copy number reported by Ghaemmaghami et al 2003
\item Cluster corresponding cluster number
\item pValue_ANOVA_test resulting p-value of an anova test
\item GOBP_terms gene ontology biological process terms
\item GOBP_names gene ontology biological process names
\item GOMF_terms gene ontology molecular function terms
\item GOMF_names gene ontology molecular function names
\item GOCC_terms gene ontology cellular component terms
\item GOC_names gene ontology cellular component names
\item Copy_number_Gal_Rep1 protein copy number based on the proteomic ruler method galactose (Gal) replicate 1
\item Copy_number_Gal_Rep2 protein copy number based on the proteomic ruler method galactose (Gal) replicate 2
\item Copy_number_Gal_Rep3 protein copy number based on the proteomic ruler method galactose (Gal) replicate 3
\item Copy_number_Glc_Rep1 protein copy number based on the proteomic ruler method glucose (Glc) replicate 1
\item Copy_number_Glc_Rep2 protein copy number based on the proteomic ruler method glucose (Glc) replicate 2
\item Copy_number_Glc_Rep3 protein copy number based on the proteomic ruler method glucose (Glc) replicate 3
\item Copy_number_Gly_Rep1 protein copy number based on the proteomic ruler method glycerol (Gly) replicate 1
\item Copy_number_Gly_Rep2 protein copy number based on the proteomic ruler method glycerol (Gly) replicate 2
\item Copy_number_Gly_Rep3 protein copy number based on the proteomic ruler method glycerol (Gly) replicate 3
\item Log2_copy_number_Gal_Rep1 protein copy number based on the proteomic ruler method galactose (Gal) replicate 1, log2 transformed
\item Log2_copy_number_Gal_Rep2 protein copy number based on the proteomic ruler method galactose (Gal) replicate 2, log2 transformed
\item Log2_copy_number_Gal_Rep3 protein copy number based on the proteomic ruler method galactose (Gal) replicate 3, log2 transformed
\item Log2_copy_number_Glc_Rep1 protein copy number based on the proteomic ruler method glucose (Glc) replicate 1, log2 transformed
\item Log2_copy_number_Glc_Rep2 protein copy number based on the proteomic ruler method glucose (Glc) replicate 2, log2 transformed
\item Log2_copy_number_Glc_Rep3 protein copy number based on the proteomic ruler method glucose (Glc) replicate 3, log2 transformed
\item Log2_copy_number_Gly_Rep1 protein copy number based on the proteomic ruler method glycerol (Gly) replicate 1, log2 transformed
\item Log2_copy_number_Gly_Rep2 protein copy number based on the proteomic ruler method glycerol (Gly) replicate 2, log2 transformed
\item Log2_copy_number_Gly_Rep3 protein copy number based on the proteomic ruler method glycerol (Gly) replicate 3, log2 transformed
\item MS_Intensity_Gal_Rep1 summed up eXtracted Ion Current (XIC) of all isotopic clusters associated with the identified AA sequence, galactose (Gal) replicate 1
\item MS_Intensity_Gal_Rep2 summed up eXtracted Ion Current (XIC) of all isotopic clusters associated with the identified AA sequence, galactose (Gal) replicate 2
\item MS_Intensity_Gal_Rep3 summed up eXtracted Ion Current (XIC) of all isotopic clusters associated with the identified AA sequence, galactose (Gal) replicate 3
\item MS_Intensity_Glc_Rep1 summed up eXtracted Ion Current (XIC) of all isotopic clusters associated with the identified AA sequence, glucose (Glc) replicate 1
\item MS_Intensity_Glc_Rep2 summed up eXtracted Ion Current (XIC) of all isotopic clusters associated with the identified AA sequence, glucose (Glc) replicate 2
\item MS_Intensity_Glc_Rep3 summed up eXtracted Ion Current (XIC) of all isotopic clusters associated with the identified AA sequence, glucose (Glc) replicate 3
\item MS_Intensity_Gly_Rep1 summed up eXtracted Ion Current (XIC) of all isotopic clusters associated with the identified AA sequence, glycerol (Gly) replicate 1
\item MS_Intensity_Gly_Rep2 summed up eXtracted Ion Current (XIC) of all isotopic clusters associated with the identified AA sequence, glycerol (Gly) replicate 2
\item MS_Intensity_Gly_Rep3 summed up eXtracted Ion Current (XIC) of all isotopic clusters associated with the identified AA sequence, glycerol (Gly) replicate 3
\item mean_log2_copy_number_Gal protein mean copy number grown on galactose, log2 transformed
\item mean_log2_copy_number_Gly protein mean copy number grown on glucose, log2 transformed
\item mean_log2_copy_number_Glc protein mean copy number grown on glycerol, log2 transformed
\item Gal_div_Glc ratio of mean log2 ratio galactose vs mean log2 ratio glucose
\item Gly_div_Glc ratio of mean log2 ratio glycerol vs mean log2 ratio glucose
}
}
\examples{
data(cellr_morgenstern_et_al_2017)
str(cellr_morgenstern_et_al_2017)
}
\references{
Morgenstern, M., et al. Definition of a High-Confidence Mitochondrial Proteome at Quantitative Scale. Cell reports 2017;19(13):2836-2852.
}
\keyword{datasets}
