% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustpro.R
\name{clustpro}
\alias{clustpro}
\title{title: "clustpro - a cluster analysis package"
author: "Christian Peikert and Muhammad Numair Mansur"
version: 0.01b
date: "19. Januar 2017"
<Add Description>}
\usage{
clustpro(matrix, method = "kmeans", min_k = 2, max_k = 10, fixed_k = -1,
  perform_clustering = TRUE, cluster_ids = NULL, rows = TRUE,
  cols = TRUE, tooltip = NULL, save_widget = TRUE, color_legend = NULL,
  width = NULL, height = NULL, export_dir = NA, export_type = "svg")
}
\description{
title: "clustpro - a cluster analysis package"
author: "Christian Peikert and Muhammad Numair Mansur"
version: 0.01b
date: "19. Januar 2017"
<Add Description>
}

