% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustpro.R
\name{clustpro_example}
\alias{clustpro_example}
\title{Example function 01}
\usage{
clustpro_example()
}
\value{
see clustpro() function output
}
\description{
A theoretical proteomics dataset composed of 1000 human proteins (UniProt accessionn umbers) and random choosen values will be used to call the clustpro() main function.
}
