% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustpro.R
\name{color_spectrum_unqiue_breaks}
\alias{color_spectrum_unqiue_breaks}
\title{###}
\usage{
color_spectrum_unqiue_breaks(values, color_spect, shift_factor = 1e-10)
}
\arguments{
\item{values}{should be a list which define the breaks of the color space. color_spect should be a list of color. Keep in mean that there must be 1 more board in the vaules list than color in color_spect.}
}
\description{
#' Function to to define the color spectrum for heatmaps
#'
#' This function allows you to define the color spectrum for heatmaps.
#' @param values should be a list which define the breaks of the color space. color_spect should be a list of color. Keep in mean that there must be 1 more board in the vaules list than color in color_spect.
#' @keywords color spectrum heatmaps
#' @export
#' @examples
#' color_spectrum()
}
\details{
color_spectrum <- function(values, color_spect){
  index <- 1
  palette <- c()
  colors <- c()
  while(index<length(values)){
    colors <- c(colors,seq(values[index],values[index+1],length=100))
    index <- index +1
  }
  palette <- colorRampPalette(color_spect)(n = ((length(values)-1)*100)-1)
  ## standard global variables can define with <<-. However, in this case the variable were blocked by the enviroment so I have to use the
  ## assign function. The first argument is the global variable name, second the local variable in function and the last the target enviroment
  assign('colors', colors, envir = .GlobalEnv)
  assign('palette', palette, envir = .GlobalEnv)
}

Function to to define the color spectrum for heatmaps

This function allows you to define the color spectrum for heatmaps.
}
\examples{
color_spectrum()
}
\keyword{color}
\keyword{heatmaps}
\keyword{spectrum}

