% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustpro.R
\name{createTooltipList}
\alias{createTooltipList}
\title{Function to create the tooltip list from a data.frame}
\usage{
createTooltipList(data, selected_columns = NULL)
}
\arguments{
\item{data}{numeric data.frame}

\item{selected_columns}{null or list of selected columns of data}
}
\description{
This function allows you convert selected columns of a data.frame into a list
}
\examples{
createTooltipList(data = mtcars, selected_columns =  c("cyl","mpg"))
}
\keyword{tooltip}
