% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustpro.R
\name{distributions_histograms}
\alias{distributions_histograms}
\title{distributions_histograms}
\usage{
distributions_histograms(matrix, export_type)
}
\arguments{
\item{matrix}{numeric data.frame}

\item{export_type}{character; type of exported graphics, tested for tiff and svg}
}
\description{
.................
}
\examples{
\dontrun{
distributions_histograms(mtcars[1:3],'tiff')
}
}
