% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustpro.R
\name{findk_kmeans}
\alias{findk_kmeans}
\title{findk_kmeans}
\usage{
findk_kmeans(matrix, k, seed = NULL)
}
\arguments{
\item{matrix}{numeric data.frame}

\item{k}{number of clusters, k}

\item{seed}{natural number, useful for creating simulations or random objects that can be reproduced}
}
\description{
.................
}
\examples{
findk_kmeans()
}

