% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustpro.R
\name{get_best_k}
\alias{get_best_k}
\title{Get best k}
\usage{
get_best_k(matrix, min_k, max_k, method, cores = 1, seed = NULL)
}
\arguments{
\item{matrix}{numeric data.frame}

\item{min_k, max_k, fixed_k}{number of clusters, k; if fixed_k is a natural number > 0, k is set to fixed_k. Otherwise the a function is called to find the optimal k for this data in the range defined by the minimum and maximal k.}

\item{method}{character; one of the following cluster methods: kmeans, cmeans}

\item{cores}{natural number, number of nodes/cores used for parallelisation}

\item{seed}{natural number, useful for creating simulations or random objects that can be reproduced}
}
\description{
.................
}
\examples{
get_best_k()
}

