% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustpro.R
\name{get_color}
\alias{get_color}
\title{Function to get color}
\usage{
get_color(x, ticks, colors)
}
\arguments{
\item{x}{numeric; value in range of the ticks}

\item{ticks}{numeric vector breaks/ticks for a gradient}

\item{colors}{character vector of colors for a gradient}
}
\description{
..............
}
\examples{
\dontrun{
get_color(x = 2, ticks = seq(1,3,1), colors = c('red','yellow','green'))
}
}
