% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustpro.R
\name{initialize_graphic}
\alias{initialize_graphic}
\title{Function to initialize a graphic}
\usage{
initialize_graphic(title, type = graphic_type, ...)
}
\description{
This function allows you to initialize a graphic
}
