% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustpro.R
\name{is.Newick}
\alias{is.Newick}
\title{Function to prove if sting is in newick format}
\usage{
is.Newick(newick_string)
}
\arguments{
\item{newick_string}{string to test}
}
\description{
This function returns TRUE is string is in newick format, FALSE otherwise
}
\examples{
is.Newick(newick_string = "(1:0.1,(2:0.2,(3:0.3,4:0.4):0.5)):0.7")
}
\keyword{newickformat}
