% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runExample.R
\docType{data}
\name{proteomics_data}
\alias{proteomics_data}
\title{Proteomic dataset}
\format{A data frame with  2590 rows and 6 variables}
\usage{
data(proteomics_data)
}
\description{
A dataset containing ......
The variables are as follows:
}
\details{
\itemize{
  \item uniProtID. id
  \item MTP_MT. mean ratio of MTP MT
  \item MT_MB. mean ratio of MT_MB
  \item MTP_MB. mean ratio of MTP_MB
  \item geneNames. gene names of proteins
  \item definition. definition of proteins
}
}
\references{
Warscheid Lab, Freiburg University
}
\keyword{datasets}

