% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cellr_morgenstern_et_al_2017}
\alias{cellr_morgenstern_et_al_2017}
\title{Morgenstern et al., Cell Reports 2017}
\format{A data frame with  1576 rows and 23 variables}
\usage{
data(cellr_morgenstern_et_al_2017)
}
\description{
Project description: Mitochondria perform central functions in cellular bioenergetics, metabolism and signaling and their malfunction has been linked to numerous diseases.
The available studies cover only part of the mitochondrial proteome and a separation of core mitochondrial proteins from associated fractions has not been achieved.
We developed an integrative, quantitative MS-based experimental approach to define the high confidence proteome of yeast mitochondria and to identify new mitochondrial proteins.
The analysis includes protein abundance under fermentable and non-fermentable growth, submitochondrial localization, single-protein analysis and subcellular classification of mitochondria-associated fractions.
We identified novel mitochondrial interactors of respiratory chain supercomplexes, ATP synthase, AAA proteases, the mitochondrial contact site and cristae organizing system (MICOS) and coenzyme Q biosynthesis cluster as well as new mitochondrial proteins with dual cellular localization.
The integrative proteome provides a high confidence source for the characterization of physiological and pathophysiological functions of mitochondria and their integration into the cellular environment.
Data: This data are a summary of proteome-wide absolute quantification of proteins under fermentable and non-fermentable growth conditions based on the 'Proteomic Ruler' approach.
Raw MS data and complete MaxQuant results files are available via ProteomeXchange with identifier PXD006146.
The variables are as follows:
}
\details{
\itemize{
\item Systematic_names ToDo
\item Uniprot_ID ToDo
\item Gene_names ToDo
\item Protein_Description ToDo
\item Moleular_weight_Da ToDo
\item High_confidence_mito_proteome ToDo
\item Sequence_coverage ToDo
\item Mean_copy_number_Glucose ToDo
\item Mean_copy_number_Galactose ToDo
\item Mean_copy_number_Glycerol ToDo
\item Absolute_quantification_accuracy ToDo
\item Copy_number_Kulak_et_al_2014 ToDo
\item Copy_number_Chong_et_al_2015 ToDo
\item Copy_number_Ghaemmaghami_et_al_2003 ToDo
\item Cluster ToDo
\item pValue_ANOVA_test ToDo
\item GOBP_terms ToDo
\item GOBP_names ToDo
\item GOMF_terms ToDo
\item GOMF_names ToDo
\item GOCC_terms ToDo
\item GOC_names ToDo
\item Copy_number_Gal_Rep1 ToDo
\item Copy_number_Gal_Rep2 ToDo
\item Copy_number_Gal_Rep3 ToDo
\item Copy_number_Glc_Rep1 ToDo
\item Copy_number_Glc_Rep2 ToDo
\item Copy_number_Glc_Rep3 ToDo
\item Copy_number_Gly_Rep1 ToDo
\item Copy_number_Gly_Rep2 ToDo
\item Copy_number_Gly_Rep3 ToDo
\item Log2_copy_number_Gal_Rep1 ToDo
\item Log2_copy_number_Gal_Rep2 ToDo
\item Log2_copy_number_Gal_Rep3 ToDo
\item Log2_copy_number_Glc_Rep1 ToDo
\item Log2_copy_number_Glc_Rep2 ToDo
\item Log2_copy_number_Glc_Rep3 ToDo
\item Log2_copy_number_Gly_Rep1 ToDo
\item Log2_copy_number_Gly_Rep2 ToDo
\item Log2_copy_number_Gly_Rep3 ToDo
\item MS_Intensity_Gal_Rep1 ToDo
\item MS_Intensity_Gal_Rep2 ToDo
\item MS_Intensity_Gal_Rep3 ToDo
\item MS_Intensity_Glc_Rep1 ToDo
\item MS_Intensity_Glc_Rep2 ToDo
\item MS_Intensity_Glc_Rep3 ToDo
\item MS_Intensity_Gly_Rep1 ToDo
\item MS_Intensity_Gly_Rep2 ToDo
\item MS_Intensity_Gly_Rep3 ToDo
\item mean_log2_copy_number_Gal ToDo
\item mean_log2_copy_number_Gly ToDo
\item mean_log2_copy_number_Glc ToDo
\item Gal_div_Glc ToDo
\item Gly_div_Glc ToDo
}
}
\references{
Morgenstern, M., et al. Definition of a High-Confidence Mitochondrial Proteome at Quantitative Scale. Cell reports 2017;19(13):2836-2852.
}
\keyword{datasets}
