% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cellr_morgenstern_et_al_2017}
\alias{cellr_morgenstern_et_al_2017}
\title{Morgenstern et al., Cell Reports 2017}
\format{A data frame with  1576 rows and 23 variables}
\usage{
data(cellr_morgenstern_et_al_2017)
}
\description{
Summary of proteome-wide absolute quantification of proteins under fermentable and non-fermentable growth conditions based on the 'Proteomic Ruler' approach.
Raw MS data and complete MaxQuant results files are available via ProteomeXchange with identifier PXD006146.
The variables are as follows:
}
\details{
\itemize{
\item Systematic_names ToDo
\item Uniprot_ID ToDo
\item Gene_names ToDo
\item Protein_Description ToDo
\item Moleular_weight_Da ToDo
\item High_confidence_mito_proteome ToDo
\item Sequence_coverage ToDo
\item Mean_copy_number_Glucose ToDo
\item Mean_copy_number_Galactose ToDo
\item Mean_copy_number_Glycerol ToDo
\item Absolute_quantification_accuracy ToDo
\item Copy_number_Kulak_et_al_2014 ToDo
\item Copy_number_Chong_et_al_2015 ToDo
\item Copy_number_Ghaemmaghami_et_al_2003 ToDo
\item Cluster ToDo
\item pValue_ANOVA_test ToDo
\item GOBP_terms ToDo
\item GOBP_names ToDo
\item GOMF_terms ToDo
\item GOMF_names ToDo
\item GOCC_terms ToDo
\item GOC_names ToDo
\item Copy_number_Gal_Rep1 ToDo
\item Copy_number_Gal_Rep2 ToDo
\item Copy_number_Gal_Rep3 ToDo
\item Copy_number_Glc_Rep1 ToDo
\item Copy_number_Glc_Rep2 ToDo
\item Copy_number_Glc_Rep3 ToDo
\item Copy_number_Gly_Rep1 ToDo
\item Copy_number_Gly_Rep2 ToDo
\item Copy_number_Gly_Rep3 ToDo
\item Log2_copy_number_Gal_Rep1 ToDo
\item Log2_copy_number_Gal_Rep2 ToDo
\item Log2_copy_number_Gal_Rep3 ToDo
\item Log2_copy_number_Glc_Rep1 ToDo
\item Log2_copy_number_Glc_Rep2 ToDo
\item Log2_copy_number_Glc_Rep3 ToDo
\item Log2_copy_number_Gly_Rep1 ToDo
\item Log2_copy_number_Gly_Rep2 ToDo
\item Log2_copy_number_Gly_Rep3 ToDo
\item MS_Intensity_Gal_Rep1 ToDo
\item MS_Intensity_Gal_Rep2 ToDo
\item MS_Intensity_Gal_Rep3 ToDo
\item MS_Intensity_Glc_Rep1 ToDo
\item MS_Intensity_Glc_Rep2 ToDo
\item MS_Intensity_Glc_Rep3 ToDo
\item MS_Intensity_Gly_Rep1 ToDo
\item MS_Intensity_Gly_Rep2 ToDo
\item MS_Intensity_Gly_Rep3 ToDo
\item mean_log2_copy_number_Gal ToDo
\item mean_log2_copy_number_Gly ToDo
\item mean_log2_copy_number_Glc ToDo
\item Gal_div_Glc ToDo
\item Gly_div_Glc ToDo
}
}
\references{
Warscheid Lab, Freiburg University
}
\keyword{datasets}
