% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustpro.R
\name{clustering}
\alias{clustering}
\title{Clustering}
\usage{
clustering(matrix, min_k = 2, max_k = 100, fixed_k = -1,
  method = "kmeans", cores = 2, seed = NULL, graphics_export = FALSE)
}
\description{
This function allows you to initialize a graphic
}

