% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustpro.R
\name{clustpro}
\alias{clustpro}
\title{<Add Title>}
\usage{
clustpro(matrix, method = "kmeans", min_k = 2, max_k = 10, fixed_k = -1,
  tooltip = NULL, save_widget = TRUE, width = NULL, height = NULL)
}
\description{
<Add Description>
}

