% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustpro.R
\name{clustpro}
\alias{clustpro}
\title{Clustpro main function}
\usage{
clustpro(matrix, method = "kmeans", hclust_method = "ward.D2", min_k = 2,
  max_k = 10, fixed_k = NULL, perform_clustering = TRUE,
  simplify_clustering = FALSE, clusterVector = NULL, rows = TRUE,
  cols = TRUE, tooltip = NULL, save_widget = TRUE, show_legend = FALSE,
  color_legend = NULL, width = NULL, height = NULL,
  export_graphics = FALSE, export_dir = NULL, export_type = "svg",
  seed = NULL, random_seeds = NULL, cores = 2, useShiny = TRUE)
}
\arguments{
\item{matrix}{numeric data.frame}

\item{method}{character; one of the following cluster methods: kmeans, cmeans}

\item{hclust_method}{character; one of the following cluster methods: "ward.D", "ward.D2", "single", "complete", "average" (= UPGMA), "mcquitty" (= WPGMA), "median" (= WPGMC) or "centroid" (= UPGMC)}

\item{min_k, max_k, fixed_k}{number of clusters, k; if fixed_k is a natural number > 0, k is set to fixed_k. Otherwise the a function is called to find the optimal k for this data in the range defined by the minimum and maximal k.}

\item{perform_clustering}{boolean; if true a clustering is performed}

\item{simplify_clustering}{boolean; if true a each cluster is represended by its mean values over all within the cluster itself.}

\item{clusterVector}{list or vector of natural number; for each row a cluster has to been given}

\item{rows, cols}{boolean; if true a hierarchical clustering for row / columns of the clustered matrix is performed}

\item{tooltip}{list of lists; list of lists containing information of each row e.g. a list for name, description.}

\item{save_widget}{boolean; if to TRUE html widget is saved as html page}

\item{show_legend}{boolean; if TRUE color legend is shown}

\item{color_legend}{list of lists; required lists: ticks and colors; ticks is a list of n breaks for the heatmap; colors is a list of n-1 colors; in addition a list for of position for shown labels can added  call label_position}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{export_graphics}{boolean; if TRUE grephics are exported}

\item{export_dir}{character; storage directory}

\item{export_type}{character; type of exported graphics, tested for tif and svg}

\item{seed}{natural number, useful for creating simulations or random objects that can be reproduced}

\item{random_seeds}{null or natural number, if null seed is used for the determination of the best k, otherwise the number of random_seeds seeds will be consider}

\item{cores}{natural number, number of nodes/cores used for parallelisation}

\item{useShiny}{if TRUE html widget is usable for shiny apps, otherwise clustering is returned to R}
}
\value{
see clustpro() function output
}
\description{
This function is used to start the clustering and visualisation process.
}
