% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustpro.R
\name{color_spectrum_unqiue_breaks}
\alias{color_spectrum_unqiue_breaks}
\title{Function to to define the color spectrum for heatmaps}
\usage{
color_spectrum_unqiue_breaks(values, color_spect, shift_factor = 1e-10)
}
\arguments{
\item{values}{should be a list which define the breaks of the color space. color_spect should be a list of color. Keep in mean that there must be 1 more board in the vaules list than color in color_spect.}
}
\description{
This function allows you to define the color spectrum for heatmaps.
}
\examples{
color_spectrum()
}
\keyword{color}
\keyword{heatmaps}
\keyword{spectrum}

