% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustpro.R
\name{get_best_k}
\alias{get_best_k}
\title{Get best k}
\usage{
get_best_k(matrix, min_k = 2, max_k = 10, method = "kmeans", cores = 1,
  seed = NULL)
}
\arguments{
\item{matrix}{numeric data.frame}

\item{min_k, max_k}{number of clusters, k; if the  function is called its tries to find the optimal k for this data in the range defined by the minimum and maximal k.}

\item{method}{character; one of the following cluster methods: kmeans, cmeans}

\item{cores}{natural number, number of nodes/cores used for parallelisation}

\item{seed}{natural number, useful for creating simulations or random objects that can be reproduced}
}
\description{
.................
}
