% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustpro.R
\name{initialize_graphic}
\alias{initialize_graphic}
\title{Function to initialize a graphic}
\usage{
initialize_graphic(filename, type = "tif", ...)
}
\arguments{
\item{filename}{character; filename for saving}

\item{type}{character; type of exported graphics, tif, svg or pdf}

\item{...}{further parameters}
}
\description{
This function allows you to initialize a graphic
}
