% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustpro.R
\name{is.Newick}
\alias{is.Newick}
\title{Function to prove if sting is in newick format}
\usage{
is.Newick(newick_string)
}
\arguments{
\item{x}{numeric data.frame}
}
\description{
This function returns TRUE is string is in newick format, FALSE otherwise
}
\keyword{newickformat}
