% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mcp_reimann_et_al_2017}
\alias{mcp_reimann_et_al_2017}
\title{Reimann et al., Molecular & Cellular Proteomics 2017}
\format{A data frame with  2588 rows and 23 variables}
\usage{
data(mcp_reimann_et_al_2017)
}
\description{
A dataset containing ......
The variables are as follows:
}
\details{
\itemize{
\item Uniprot Uniprot accession numbers of all proteins assigned to the respective protein group
\item First_ID First ID of Protein group(s) from MaxQuant output file proteinGroups.txt
\item Gene_names Gene names of all proteins assigned to the respective group
\item Protein_names Protein names of all proteins assigned to the respective group
\item Main_cluster Assignment of proteins to the 4 main cluster after k-means clustering
\item Cluster Assignment of proteins to the distinct cluster after k-means clustering
\item Mol_weight_kDa Molecular weight in kDa
\item Number_of_proteins Number of proteins assigned to the respective protein group
\item Peptides Number of identified peptides assigned to the respective protein group
\item Unique_peptides Number of peptide sequences identified across all replicates of this dataset that are unique for the respective protein group
\item Sequence_coverage Percentage of the sequence that is covered by the identified peptides of the best protein sequence contained in the group
\item Mean_log10_ratio_MTP_vs_MT The mean of the log10 transformed protein ratios of paced myotubes MTP to unpaced myotubes MT
\item Mean_log10_ratio_MT_vs_MB The mean of the log10 transformed protein ratios of unpaced myotubes MT to myoblasts MB
\item Mean_log10_ratio_MTP_vs_MB The mean of the log10 transformed protein ratios of paced myotubes MTP to myoblasts MB
\item Minus_log10_p_value_MTP_vs_MB The negative log10 transformed p-value of paced myotubes MTP to unpaced myotubes MT
\item Minus_log10_p_value_MT_vs_MB The negative log10 transformed p-value of unpaced myotubes MT to myoblasts MB
\item Minus_log10_p_value_MTP_vs_MT The negative log10 transformed p-value of paced myotubes MTP to myoblasts MB
\item log10_ratio_MTP_vs_MT The log10 transformed protein ratios of paced myotubes MTP to unpaced myotubes MT for each replicate
\item log10_ratio_MT_vs_MB The log10 transformed protein ratios of unpaced myotubes MT to myoblasts MB for each replicate
\item log10_ratio_MTP_vs_MB The log10 transformed protein ratios of paced myotubes MTP to myoblasts MB for each replicate
}
}
\references{
Warscheid Lab, Freiburg University
}
\keyword{datasets}
