% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mcp_reimann_et_al_2017}
\alias{mcp_reimann_et_al_2017}
\title{Reimann et al., Molecular & Cellular Proteomics 2017}
\format{A data frame with 2588 rows and 17 variables}
\usage{
data(mcp_reimann_et_al_2017)
}
\description{
Project description: The Z-disc is a protein-rich structure critically important for development and integrity of myofibrils in muscle cells.
Changes in protein abundance upon myoblast differentiation into myotubes and upon formation of contracting myotubes generated by electrical pulse stimulation were monitored in a quantitative proteomics study.
Data: Excerpts from the MaxQuant output files "proteinGroups.txt" (spreadsheet proteins); details for protein/protein group identifications are given.
Identifications referring to reverse hits and potential contaminants have been removed. Raw MS data and complete MaxQuant result files are available via ProteomeXchange with identifier PXD004960.
}
\details{
\itemize{
\item Uniprot Uniprot accession numbers of all proteins assigned to the respective protein group
\item First_ID First ID of Protein group(s) from MaxQuant output file proteinGroups.txt
\item Gene_names Gene names of all proteins assigned to the respective group
\item Protein_names Protein names of all proteins assigned to the respective group
\item Main_cluster Assignment of proteins to one of the 4 main clusters after k-means clustering
\item Cluster Assignment of proteins to the distinct cluster after k-means clustering
\item Mol_weight_kDa Molecular weight in kDa
\item Number_of_proteins Number of proteins assigned to the respective protein group
\item Peptides Number of identified peptides assigned to the respective protein group
\item Unique_peptides Number of peptide sequences identified across all replicates of this dataset that are unique for the respective protein group
\item Sequence_coverage Percentage of the sequence that is covered by the identified peptides of the best protein sequence contained in the group
\item Mean_log10_ratio_MTP_vs_MT The mean of the log10 transformed protein ratios of paced myotubes MTP to unpaced myotubes MT
\item Mean_log10_ratio_MT_vs_MB The mean of the log10 transformed protein ratios of unpaced myotubes MT to myoblasts MB
\item Mean_log10_ratio_MTP_vs_MB The mean of the log10 transformed protein ratios of paced myotubes MTP to myoblasts MB
\item Minus_log10_p_value_MTP_vs_MB The negative log10 transformed p-value of paced myotubes MTP to unpaced myotubes MT
\item Minus_log10_p_value_MT_vs_MB The negative log10 transformed p-value of unpaced myotubes MT to myoblasts MB
\item Minus_log10_p_value_MTP_vs_MT The negative log10 transformed p-value of paced myotubes MTP to myoblasts MB
}
}
\examples{
data(mcp_reimann_et_al_2017)
str(mcp_reimann_et_al_2017)
}
\references{
Reimann, L., et al. Myofibrillar Z-discs Are a Protein Phosphorylation Hot Spot with Protein Kinase C (PKCalpha) Modulating Protein Dynamics. Molecular & Cellular Proteomics 2017;16(3):346-367.
}
\keyword{datasets}
