% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustpro.R
\name{setHeatmapColors}
\alias{setHeatmapColors}
\title{Function to set heatmap color}
\usage{
setHeatmapColors(data, color_list = c("red", "yellow", "green"), intervals,
  auto = FALSE)
}
\arguments{
\item{data}{todo}

\item{color_list}{todo}

\item{intervals}{todo}

\item{auto}{todo}
}
\description{
This function allows you to define the color spectrum for heatmaps.
}
\examples{
setHeatmapColors()
}
\keyword{color}
\keyword{heatmaps}
\keyword{spectrum}

