% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustpro.R
\name{setHeatmapColors}
\alias{setHeatmapColors}
\title{Function to set heatmap color}
\usage{
setHeatmapColors(data, color_list = c("red", "yellow", "green"), intervals,
  auto = FALSE)
}
\arguments{
\item{data}{numeric data.frame, list or vector}

\item{color_list}{list of colors}

\item{intervals}{list of numerical vaules which define the breaks of the color space}

\item{auto}{boolean, if TRUE automatical gerneates ticks for the dataset}
}
\description{
This function allows you to define the color spectrum for heatmaps.
}
\keyword{color}
\keyword{heatmaps}
\keyword{spectrum}
